[FLINK-21322][coordination] Make certain ExecutionGraph-logic re-usable
